# api/openapi.json
{
  "components": {
    "schemas": {
...
      "EventDetail": {
        "properties": {
          "camera_id": {
            "title": "Camera Id",
            "type": "string"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "plate_argmax": {
            "title": "Plate Argmax",
            "type": "string"
          },
          "plate_confidence": {
            "title": "Plate Confidence",
            "type": "number"
          },
          "plate_posterior": {
            "items": {
              "$ref": "#/components/schemas/SlotPosteriorOut"
            },
            "title": "Plate Posterior",
            "type": "array"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "trajectory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trajectory Id"
          },
          "vehicle_type": {
            "title": "Vehicle Type",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "camera_id",
          "timestamp",
          "plate_argmax",
          "plate_confidence",
          "color",
          "vehicle_type",
          "trajectory_id",
          "plate_posterior"
        ],
        "title": "EventDetail",
        "type": "object"
      },
      "EventSummary": {
        "properties": {
          "camera_id": {
            "title": "Camera Id",
            "type": "string"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "plate_argmax": {
            "title": "Plate Argmax",
            "type": "string"
          },
          "plate_confidence": {
            "title": "Plate Confidence",
            "type": "number"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "trajectory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trajectory Id"
          },
          "vehicle_type": {
            "title": "Vehicle Type",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "camera_id",
          "timestamp",
          "plate_argmax",
          "plate_confidence",
          "color",
          "vehicle_type",
          "trajectory_id"
        ],
        "title": "EventSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinkEvidence": {
        "properties": {
          "appearance_lr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appearance Lr"
          },
          "delta_t_s": {
            "title": "Delta T S",
            "type": "number"
          },
          "expected_t_s": {
            "title": "Expected T S",
            "type": "number"
          },
          "from_event_id": {
            "title": "From Event Id",
            "type": "string"
          },
          "kinematic_lr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kinematic Lr"
          },
          "plate_lr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plate Lr"
          },
          "prior_log_odds": {
            "title": "Prior Log Odds",
            "type": "number"
          },
          "skipped_cameras": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Cameras",
            "type": "array"
          },
          "to_event_id": {
            "title": "To Event Id",
            "type": "string"
          },
          "total_log_odds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Log Odds"
          }
        },
        "required": [
          "from_event_id",
          "to_event_id",
          "plate_lr",
          "appearance_lr",
          "kinematic_lr",
          "prior_log_odds",
          "total_log_odds",
          "delta_t_s",
          "expected_t_s",
          "skipped_cameras"
        ],
        "title": "LinkEvidence",
        "type": "object"
      },
...
      "Page_EventSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[EventSummary]",
        "type": "object"
      },
      "Page_TrajectorySummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrajectorySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[TrajectorySummary]",
        "type": "object"
      },
      "PathPoint": {
        "properties": {
          "camera_id": {
            "title": "Camera Id",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "camera_id",
          "lat",
          "lon",
          "timestamp"
        ],
        "title": "PathPoint",
        "type": "object"
      },
      "PlateConsensus": {
        "properties": {
          "entropy_bits": {
            "title": "Entropy Bits",
            "type": "number"
          },
          "per_slot": {
            "items": {
              "items": {
                "$ref": "#/components/schemas/SlotRead"
              },
              "type": "array"
            },
            "title": "Per Slot",
            "type": "array"
          },
          "single_read_plates": {
            "items": {
              "type": "string"
            },
            "title": "Single Read Plates",
            "type": "array"
          }
        },
        "required": [
          "per_slot",
          "single_read_plates",
          "entropy_bits"
        ],
        "title": "PlateConsensus",
        "type": "object"
      },
...
      "SearchHit": {
        "properties": {
          "matched_plate": {
            "title": "Matched Plate",
            "type": "string"
          },
          "probability": {
            "title": "Probability",
            "type": "number"
          },
          "trajectory": {
            "$ref": "#/components/schemas/TrajectorySummary"
          }
        },
        "required": [
          "trajectory",
          "probability",
          "matched_plate"
        ],
        "title": "SearchHit",
        "type": "object"
      },
      "SlotPosteriorOut": {
        "properties": {
          "top": {
            "items": {
              "$ref": "#/components/schemas/SlotRead"
            },
            "title": "Top",
            "type": "array"
          },
          "unread": {
            "title": "Unread",
            "type": "boolean"
          }
        },
        "required": [
          "top",
          "unread"
        ],
        "title": "SlotPosteriorOut",
        "type": "object"
      },
      "SlotRead": {
        "properties": {
          "char": {
            "title": "Char",
            "type": "string"
          },
          "prob": {
            "title": "Prob",
            "type": "number"
          }
        },
        "required": [
          "char",
          "prob"
        ],
        "title": "SlotRead",
        "type": "object"
      },
      "TrajectoryDetail": {
        "properties": {
          "camera_sequence": {
            "items": {
              "type": "string"
            },
            "title": "Camera Sequence",
            "type": "array"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "consensus": {
            "$ref": "#/components/schemas/PlateConsensus"
          },
          "decoded_plate": {
            "title": "Decoded Plate",
            "type": "string"
          },
          "end_time": {
            "title": "End Time",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "title": "Events",
            "type": "array"
          },
          "has_alert": {
            "title": "Has Alert",
            "type": "boolean"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkEvidence"
            },
            "title": "Links",
            "type": "array"
          },
          "n_events": {
            "title": "N Events",
            "type": "integer"
          },
          "path": {
            "items": {
              "$ref": "#/components/schemas/PathPoint"
            },
            "title": "Path",
            "type": "array"
          },
          "plate_confidence": {
            "title": "Plate Confidence",
            "type": "number"
          },
          "start_time": {
            "title": "Start Time",
            "type": "string"
          },
          "trajectory_id": {
            "title": "Trajectory Id",
            "type": "string"
          },
          "vehicle_type": {
            "title": "Vehicle Type",
            "type": "string"
          }
        },
        "required": [
          "trajectory_id",
          "decoded_plate",
          "plate_confidence",
          "start_time",
          "end_time",
          "n_events",
          "camera_sequence",
          "color",
          "vehicle_type",
          "has_alert",
          "events",
          "links",
          "path",
          "consensus"
        ],
        "title": "TrajectoryDetail",
        "type": "object"
      },
      "TrajectorySummary": {
        "properties": {
          "camera_sequence": {
            "items": {
              "type": "string"
            },
            "title": "Camera Sequence",
            "type": "array"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "decoded_plate": {
            "title": "Decoded Plate",
            "type": "string"
          },
          "end_time": {
            "title": "End Time",
            "type": "string"
          },
          "has_alert": {
            "title": "Has Alert",
            "type": "boolean"
          },
          "n_events": {
            "title": "N Events",
            "type": "integer"
          },
          "plate_confidence": {
            "title": "Plate Confidence",
            "type": "number"
          },
          "start_time": {
            "title": "Start Time",
            "type": "string"
          },
          "trajectory_id": {
            "title": "Trajectory Id",
            "type": "string"
          },
          "vehicle_type": {
            "title": "Vehicle Type",
            "type": "string"
          }
        },
        "required": [
          "trajectory_id",
          "decoded_plate",
          "plate_confidence",
          "start_time",
          "end_time",
          "n_events",
          "camera_sequence",
          "color",
          "vehicle_type",
          "has_alert"
        ],
        "title": "TrajectorySummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SUTRA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/events": {
      "get": {
        "operationId": "list_events_api_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "camera_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EventSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "core"
        ]
      }
    },
    "/api/events/{event_id}": {
      "get": {
        "operationId": "get_event_api_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event",
        "tags": [
          "core"
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "color",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Color"
            }
          },
          {
            "in": "query",
            "name": "vehicle_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vehicle Type"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchHit"
                  },
                  "title": "Response Search Api Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "core"
        ]
      }
    },
    "/api/trajectories": {
      "get": {
        "operationId": "list_trajectories_api_trajectories_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "plate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plate"
            }
          },
          {
            "in": "query",
            "name": "camera_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          },
          {
            "in": "query",
            "name": "min_len",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Len"
            }
          },
          {
            "in": "query",
            "name": "has_alert",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Alert"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TrajectorySummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Trajectories",
        "tags": [
          "core"
        ]
      }
    },
    "/api/trajectories/{trajectory_id}": {
      "get": {
        "operationId": "get_trajectory_api_trajectories__trajectory_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trajectory_id",
            "required": true,
            "schema": {
              "title": "Trajectory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Trajectory",
        "tags": [
          "core"
        ]
      }
    }
  }
}